requirements.txt
art
pillow
platform
os
pandas